ahtesham april 30 update routing

+++ src/app/features/features.component.ts
@@ -112,13 +112,15 @@ export class FeaturesComponent implements AfterViewInit {
 
 
   deletefeature(x,y){
-    this.backend.deletefeature(x,y)
-    .subscribe((data)=> { 
-        console.log("data",data);
-   this.getallf();
-
-    });
-
+    // alert(x)
+    if (confirm('Are you sure to delete this record ?') == true) {
+      this.backend.deletefeature(x,y)
+      .subscribe((data)=> { 
+          console.log("data",data);
+     this.getallf();
+  
+      });
+  }
   }
